Add Option to Abandon Overly Long Keyword Searches
Tame extremely large pages which take forever to search
for keywords. Mostly due to too much content on the page.
A new optional script argument has been added, patience that
abdicates when it takes too long to process. A default setting
is set to 30 seconds incase one is not provided by user.