Add Boost Terms to Refine Sorting Order
Boost terms are basically search keywords that are
counted on the page and given a standard weight (currently 10)
in order to manipulate overall relevancy / rank of a page.

For example, a scraped page with "program", "undergrad", "academic" e.t.c
will get additional points + weight so that it is ranked better than a
similar page appearing on non-relevant section of the website such as a
blog.

A new file known as boost_terms.txt, baked into the code has been added.
Edit the file to add new boost terms.

Signed-off-by: Antony Oduor <aowino@gmail.com>